requirements.txt
livekit-agents[cartesia,silero,openai]
python-dotenv
